Extend expiry of Net.WebSocket.CloseCode histogram
The WebSocketReassembleShortMessages experiment, which uses the
Net.WebSocket.CloseCode as one of its metrics, has been extended to
beta, and will probably be extended to stable.

Extend the life of the Net.WebSocket.CloseCode histogram all the way to
December 2020 so hopefully we won't have to extend it again.

BUG=1104953

Change-Id: I6e8edc2d7204628901e870c97a9248e5fcbe99b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298867
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790398}