Report MetricKit data to UMA
As AddCount is not exposed in macros, use directly histograms
functions. As only one report is received every 24 hours, not
caching the histogram should not be an issue.

This is still behind a flag and only enabled on canary.

Bug: 1017434
Change-Id: I47680e10ec5030b1d6e68c845a5cbb1ae91f92f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283742
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790421}