views: change rect targeting to favor deepest view
Prior to this patch targeting would potentially consider the parent.
This is problematic as if there is a suitable child, then preferring
the parent likely means targeting a container view, which generally
doesn't do anything with events.

Considering the parent for distance to center is also problematic
in so far it means touches on the center of a parent are preferred
over a touch on a child, even though the center is slightly off.

TEST=none
BUG=1085694

Change-Id: I8e65bddf2a48bef514094fb29c8aed2dbbc46176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2295439
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790422}