cros: Power button animation starts from current opacity.
It used to start a new animation if the power button was pressed in the
middle of a power menu animation. It's really hard to see without slow
animations, but some users have noticed.

Fixes this by using IMMEDIATE_ANIMATE_TO_TARGET to animate from the
current opacity to the target. Also avoids recreating the contents of
the widget on each press.

Test: manual
Change-Id: I2361e431cae9bb74a37b29354b6b0825e81ca4ac
Fixed: 832975
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297746
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Min Chen <minch@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790429}