[Autofill Assistant] Allow GenericUi to request UserData.
This bridges the gap between legacy UserData and the new UserModel. For
now, only additional values can be requested, but this can be easily
extended to apply to the selected card/profile etc. as well.

All values requested in this manner will be client-side only and never
leave the device.

Bug: b/145043394
Change-Id: Iaf3fe8af7a24418eabe36393b5428752a1d20aae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308417
Reviewed-by: Marian Fechete <marianfe@google.com>
Reviewed-by: Luca Hunkeler <hluca@google.com>
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#790430}