Refactor RenderWidget state for nested main frames
Split out from https://crrev.com/c/2284613.

This CL adds knowledge of a main frame RenderWidget's state as a nested
widget. We also do some related cleanups and simplifications.

GuestViews and Portals create a RenderView and associated RenderWidget
with a RenderWidgetDelegate. However, such RenderWidgets aren't the top
level widgets in the frame tree. There are cases where we want to treat
these widgets as subframes, rather than a top-level frame, such as the
CL mentioned above.

In this CL we plumb through the view type when creating a
RenderView. This information is used by RenderWidget to determine if
it's nested. Additionally, we add code to keep this bit up-to-date in
the case where a portal is activated/deactivated, changing this status.

As a small, unrelated change, we also fold setting the
RenderWidgetDelegate into RenderWidget::InitForMainFrame to make it
clear that main-frame widgets, and only main-frame widgets, have a
non-null delegate_. This change should be a no-op.

Bug: 932669
Change-Id: I685e1d13970e9c88a19c66972bd3ce3bf4749c00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302813
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790445}