Only show display highlight while dragging
This change makes it so that a highlight around the currently selected
display is only shown once dragging of the display in the Display Layout
section begins. Previously, whenever Display settings was open and
multiple displays were attached to the device, whichever display was
currently selected in settings would be highlighted at all times.

Bug: 1070122
Change-Id: I2ab461dc15db4b63b98a54383ca8966e5083ffdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290353
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790490}