Use gesture delegate for scrim component and add test
The GESTURE_DETECTOR property for the scrim is currently unused and
doesn't correctly pass motion events back to the mediator as it was
originally intended to. This patch correctly delegates event handling
to the mediator; there is no behavior change if the GESTURE_DETECTOR
property is null or not set. A test has also been added to ensure
events can be processed by the detector if set.

Bug: 1062099
Change-Id: Ia67f8f6dfa7b16d9f74a71bbe1a73994fe8ed034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2305921
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790541}