Avoid LayoutObject::FragmentsVisualRectBoundingBox() in paint timing
We are removing the function to remove obstacle to integrate paint
with LayoutNGBlockFragments.

Now always pass image border size to painting timing detectors. This
is also more accurate than the visual rect used previously. Visual rect
includes visual overflows, borders, etc. which are not part of the
image.

For the text in FileUploadControl, now report the bounds of the text
instead of the control's visual rect which also covers the button and
visual overflows.

Bug: 1104064
Change-Id: I9974ee67c5f0a0bdc84e0d0530c12c4843b577ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2309801
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790544}