Update dependency visualization graph to be completely reactive
Previously, the users of the graph visualization component would need to
ping it whenever there was a data update. Now, the users should supply
as props the properties of pageModel that the graph should be reactive
to. Since the page is now fully reactive, the "update" buttons for the
inbound/outbound components have been removed.

The goal of this is to make it easier to add UI components that
interface with the visualization component.

Bug: 1106107
Change-Id: I0578ad655fc6d13d7cc9cb7e444920846e02438b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2310194
Commit-Queue: James Long <yjlong@google.com>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Reviewed-by: Mohamed Heikal <mheikal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790560}