[omnibox] Make AX announcement when section headers are hidden/shown
When the section headers are hidden or shown by a user, currently,
screen readers give no indication that anything has changed.

This CL fixes that and adds two new explicit announcements:
 - "Related to recent searches" section shown
 - "Related to recent searches" section hidden

They are announced when the user shows or hides a section.

It also fixes a bug in OmniboxPopupContentsView that caused some weird
reentrant-like behavior, where updating the popup bounds would cause it
to call UpdatePopupAppearance again.

This bug suppressed the announcements, so this is needed too.

Bug: 1099790
Change-Id: I96d847bc304d91a37ac306e2d9a27438d68856e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2291234
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790566}