Remove usage of RenderFrameHost in WebUIURLLoaderFactory.
It didn't really need it; it was using it as a hint to know when it can be deleted when the frame goes away. A simpler method is to delete itself when it has no more receivers, if it's used in that mode. In the other mode where it's strongly owned, ownership remains the same.

Bug: 1072817

Change-Id: I9d7f86b38650093df1c7d53964c3e6123961a16f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2306013
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790567}