Improve integration with kdialog.
Modern kdialog versions support passing in filters in the format
SelectFileDialog has them, avoiding the need for a lossy conversion
via mime types. Detect the kdialog version and make use of this.

A downside of this is that is now much more obvious that always passing
1 as "selected filter index" to select file callers is wrong. But
unfortunately kdialog does not provide a way to figure out which filter
was selected when multiple filters could match the returned file. At
least this isn't more broken than it was before.

Bug: 1108060
Change-Id: Ie39baf0dfb0664e366d7ae1574303a66ab8e2de9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2311056
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790610}