Move (Begin/End)BatchReadAccessSharedImage to GLRenderer
GLRenderer/Webview use (Begin/End)BatchReadAccessSharedImage commands
to mark begin/end frame in SharedImageBatchAccessManager. Currently it's
tied to ScopedBatchedReturnResources which issues
EndBatchReadAccessSharedImage after DrawAndSwap was done and so the
command won't be flushed till next frame which leaves SharedImage in
state where EndRead was called, but no fence was produced.

To fix it and avoid extra flush move it to GLRenderer BeginDrawingFrame
and FinishDrawingFrame, so it will be flushed with SwapBuffers call.

Change-Id: I4452213c3081a700a991349d40aa65ea00a49dbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302414
Reviewed-by: Jonathan Backer <backer@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790618}