Convert remaining top level Bind/Callbacks in chrome/browser
Converts shell_integration::DefaultWebClientWorkerCallback to a
RepeatingCallback, since it is used to pass the result of state queires.
Converts callbacks for shell_integration::win::GetIsPinnedTaskbarState
to Once callbacks, given that exactly one of them is called once.
Converts the sole caller of GetIsPinnedTaskbarState to base::BindOnce.
Converts the shell_integration::win::DefaultBrowserActionRecorder to use
a OnceClosure, wince this matches it's usage.
Updates a missed comment in process_resource_usage.h
Switches the ChromeContentBrowserClient UnitTest to use a NullCallback.
This matches the intention of the test, and can be used independent of
updating the underlying API.

Finally, with all top level usages of base::Bind/Callback cleaned up,
modifies PRESUBMIT.py to replace chrome/browser with the children that
still need to be converted. Where a directory has only one file that
needs conversion, that file is listed. Otherwise, the deepest
subdirectory under chrome/browser that has files that need to be
converted is listed.

Bug: 1007635
Change-Id: I2cbdcd37cd0da411e106c0f3b14b8f6e92d7209f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2304129
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790638}