Install system webapp directly from WebApplicationInfo
In the case of system webapps, we currently install everything from
from disk, but pretend that it's remote resources over http using
the resource mechanism and a few overrides of the usual web
downloading stack. The flow basically goes:
pwa.html -> manifest.json -> (webapp.html, icons)
We're seeing installation failures of system webapps
because of this because of renderer crashes and other things.

This new flow allows installation of system webapps directly from a
WebApplicationInfo created in C++ code instead of a manifest. We now
skip the pwa.html and manifest steps entirely, and directly parse the
icons from the resource file instead of going over the network.

This only changes the sample SWA, so a follow on CL will migrate the
other SWAs to this new flow.

Bug: 1085274
Change-Id: I53e2925f2db32f81b217d258ee924865bb36a8f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240589
Commit-Queue: Dominic Schulz <dominicschulz@google.com>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790679}