Add Observing windows when modify AppService instance.
Sometimes, during the system init phases,the windows are not added to
the observe list in AppServiceAppWindowLauncherController, so when the
window is destroyed, AppServiceAppWindowLauncherController is not
notified, and the window is not removed from AppService instance, which
might cause crash when operating the destroyed window saved in instance.

This CL modified the code to check and add the window to
AppServiceAppWindowLauncherController's |observed_windows_|, when the
window is added/modified in AppService instance, so that the window
destroy callback in AppServiceAppWindowLauncherController can catch the
window destroy information to remove the window recorder.

BUG=1011235

Change-Id: I48ee1a7fc55adc9097a53f18149f44672ac10a92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2307669
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790712}