[M85] Fix Webapp.CheckServiceWorker.Time metric (CheckOfflineCapability)
Webapp.CheckServiceWorker.Time is used for the Finch experiment
(CheckOfflineCapability), but it always records a minimum value (0.01)
because the order of a subtraction was wrong. The result of a
subtraction became a minus value and it was rounded to a minimum value.

This CL fixes the order in a subtraction.

Bug: 1106693, 965802
Change-Id: Ia0e6172a34ba1e30e4c791e5e38a230aa312eaf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303414
Commit-Queue: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#846}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}