[Autofill Assistant] Added test for image resource identifiers
We used to have a test like this, but it was disabled some time ago due
to hard-to-reproduce issues on devices with older API levels. This is
a more focused test specifically for vector graphics. I have tried this
on a variety of (emulated) devices and API levels and could not
reproduce, so I'm going ahead and submit this test.

Tested on:
- Generic phone 21
- Nexus 5x 23
- Nexus 5x 25
- Nexus 5x 26
- Pixel 3 XL 28
- Pixel 3 30

Bug: b/145043394
Change-Id: I6d27eeda3e2ef7c2e29d5364838349664a83eb49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2310991
Commit-Queue: Marian Fechete <marianfe@google.com>
Reviewed-by: Marian Fechete <marianfe@google.com>
Auto-Submit: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#790785}