[Android] Fix bookmarks and settings test for master sync decoupling
Two aspects overlooked in crrev.com/c/2309699.
- One test in BookmarkTest also needs to be updated now that the promo
in the bookmarks UI is also shown when DecoupleSyncFromAndroidMasterSync
is enabled.
- In testToggleChromeSyncFromSettings(), if the master sync is not
enabled in the beginning of the test when the feature is enabled, the
waitForLastNotificationCompleted() call after disabling the toggle hangs,
since no notification is triggered when setMasterSyncAutomatically()
doesn't actually change the state of the toggle.

Bug: 1105795
Change-Id: Iaba8da32b99bac16c6103d71f63169453752c969
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2312418
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#790822}