Move media session activation logic to MediaSessionController
This gathers some pieces of MediaSession-specific logic spread across
WebMediaPlayer implementations and moves them to MediaSession-specific code
area.  The aim is for the WMPD interface to be for conveying generic
media-related events (metadata changes, play/pause, etc.), and for
MediaSessionController to be the consumer of these events that is able to
process them to make MediaSession-specific decisions (activate/suspend
MediaSession when unmuted playback starts, etc.).

This also enables future changes like teaching MediaSessionController to
activate MediaSession upon entering Picture-in-Picture even if media is not
playing.

Bug: 1091203
Change-Id: I8b2d29c0fe40f8ef80234fe832c4fd50878c1f4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308423
Commit-Queue: Wojciech Dzierżanowski <wdzierzanowski@opera.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790831}