Rewrite hr-timestamp/input-events.html to a browser test
hr-timestamp/input-events.htmlEvent timestamp tests that the received
events on the page should be equal to the timestamp provided by the
eventSender's recorded last event. But we should deprecate eventSender in
the future, and GpuBenchmarking sends events based on the current system
time, which cannot be equal to the received events' timestamp.


Bug: 1047176
Change-Id: I225dec07b1928ae254d61e8a8384ebfc7dab75e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173421
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790868}