Reraster for page scale in nested views
Page scale is changed when the user pinch-zooms a page or via parameters
controlled by the viewport meta tag. Internally, there are two notions
of page scale. A typical page hosted in a main frame will have a
|page_scale_factor| that transforms all geometry below the the visual
viewport. In this situation, a descendant widget (e.g. an OOPIF) will
have its surface in the compositor scaled as a result of the page scale
changes in the top-level frame. However, the descendant needs to update
its content to reraster at the new scale.  We cannot change the
|page_scale_factor| inside it as that would scale all the geometry
inside (the output of which is already scaled in the compositor). To
solve this, nested frame widgets use an |external_page_scale_factor|.
This value is used solely to affect the scale used for raster without
applying any transforms to its layers.

However, some widgets are in a state of being both a main frame and also
nested inside a frame hierarchy. Portals and GuestViews(*) are two such
widgets.  Because it is a main frame and runs on Android, a portaled
page may require scaling due to constraints set in the viewport <meta>
tag; this scaling must apply to the |page_scale_factor|. That is, we
want it to apply independently of the scale in the top-level frame and
remain when the portal is activated to become top-level. However, prior
to the portal being activated, we do also need to reraster its contents
based on the page scale in the embedder.

This CL does two things. First, it ensures that nested main frame
widgets, like portals and GuestViews, participate in the
|external_page_scale_factor| propagation from the main frame, as an
OOPIF would. Alone, this change breaks assumptions about the
|external_page_scale_factor| and |page_scale_factor| never being used in
tandem. Thus, we also update PictureLayerImpl to combine these two
values for raster and update DCHECKs in various places to make this a
new possibility.

(*) GuestViews also have this property but aren't really relevant in
this case as they do not support browser native pinch-zoom or scaling.
However, this fix applies to GuestViews as well, should they ever stop
suppressing pinch-zoom.

Bug: 932669
Change-Id: I0be06f577c634f3a752a8f8c2439a284f94472d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284613
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790869}