[FlexNG] Image fixes for 85 branch 4183
Both of the below cherry picks are needed to fix an image issue.


[FlexNG] Fix % max-height definiteness

A replaced element's % max-height was being ignored when it was the
child of a flex item that acquired its definite height by stretching.

This was flaky though, it only broke when the image was loaded before
the flexbox had laid out. If the flexbox was laid out, _then_ the image
loaded and hence caused a relayout, the image would obey the
definiteness that was cached from the flex item's first layout.

Fixed: 1092934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287370
Commit-Queue: David Grogan <dgrogan@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787095}

[FlexNG] Use block % resolution size when finding image preferred widths

Flex uses image preferred widths as flex-basis. When an image has
max-height:blah%, that can affect its preferred width. FlexNG wasn't
passing the percent resolution height to legacy for this calculation,
and the % max-height was being treated as auto.

ComputeMinAndMaxContentContribution was already doing this for replaced
elements, so piggy-back on that code.

Bug: 1107604
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2309048
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790439}
Change-Id: I76de49a48a1d046ea970b01eabaf4de6d9c47277
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2311309
Reviewed-by: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#854}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}