[FlexNG] Image fixes for 84 branch 4147
Both of the below cherry picks are needed to fix an image issue. A
trivial change was needed to make them compile on the 84 branch.


[FlexNG] Fix % max-height definiteness

A replaced element's % max-height was being ignored when it was the
child of a flex item that acquired its definite height by stretching.

This was flaky though, it only broke when the image was loaded before
the flexbox had laid out. If the flexbox was laid out, _then_ the image
loaded and hence caused a relayout, the image would obey the
definiteness that was cached from the flex item's first layout.

Fixed: 1092934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287370
Commit-Queue: David Grogan <dgrogan@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787095}

[FlexNG] Use block % resolution size when finding image preferred widths

Flex uses image preferred widths as flex-basis. When an image has
max-height:blah%, that can affect its preferred width. FlexNG wasn't
passing the percent resolution height to legacy for this calculation,
and the % max-height was being treated as auto.

ComputeMinAndMaxContentContribution was already doing this for replaced
elements, so piggy-back on that code.

Bug: 1107604
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2309048
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790439}

Bug: 1107428
Change-Id: I8383382c4640b608dbee09af7754f33a838ebd90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2311072
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/branch-heads/4147@{#924}
Cr-Branched-From: 16307825352720ae04d898f37efa5449ad68b606-refs/heads/master@{#768962}