[remoting host] Enable client to request ICE restart.
If SDP re-negotiation is needed (for example, to configure a different
video/audo codec), ICE needs to be restarted. But the client cannot do
this directly, because the host is the end that sends the initial SDP
offer.

This CL adds a control message that the client can use to request ICE
restart from the host. This is only implemented for WebRTC connections.

Bug: 706220
Change-Id: I44a1d7811375e4ffbc714a4cc0a8f58a75524a3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2309218
Reviewed-by: Joe Downing <joedow@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790901}