PDFViewerUpdate: Add some basic tests for the new toolbar
Adding tests for zoom, fit, and rotate buttons.

Also fixing a small bug, where forceFit() was setting the icon to the
opposite of the desired type in the new toolbar (reproduce by opening
a PDF with #view=fit appended to the URL, with the new and old UI).

Bug: 1101146
Change-Id: I86b1e0b7cdb1e5a04cb10552c904e0d58cd6f20f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2310898
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790977}