Require that the correct MatchResult::Finish*-calls take place
If the calls to MatchResult::Finish are not correct, we can end up
rules that appear to exist in the wrong origin.

In CollectPseudoRulesForElement, there is an option _not_ match
UA/User rules, but it didn't call the Finish* functions for
skipped origins.

In PseudoStyleForElementInternal, there is a call to
FinishAddingAuthorRulesForTreeScope with _may_ take place without the
other finishers being called first. Since no further rules are added
anyway, we can simply remove this call.

Change-Id: I81e54b267eba6098d9037d2d403047bc3da25dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2310338
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790995}