[Extensions] Don't show the settings overridden dialog for same-domains
An extension controlling settings, like search provider, could
potentially have the same settings as the user previously set. For
instance, a user could:
1) set search provider to example.com
2) install the example.com search extension

In this case, we shouldn't display the settings overridden dialog,
because there's no effective change.

Add logic to not display the dialog if the fallback search provider
is not extension provided, and matches the origin of the new
provider. If the previous provider was an extension, we'll still
show the dialog - this simplifies the implementation, and can also
be useful to the user because the controlling extension changed.

Add tests for the same.

Bug: 940923

Change-Id: I3ab9ba91d8bb473e44892a6bcc9d746d8472a465
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2313997
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791001}