Invalidate EDID blobs that are empty.
In some cases, when the kernel fails to read/parse an EDID the EDID
property is present but empty. We should treat an empty EDID as a
failure case and not attempt to parse, which results in inconsistent
default values.

Ex. We expect an invalid EDID to have a kInvalidProductCode. Currently,
an empty EDID blob will generate a product code of 0.

Change-Id: Ifd3f993c10e0bb728412028b357bbba0f6631146
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219518
Commit-Queue: Sasha McIntosh <sashamcintosh@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791004}