Adjust avatar profile size to avoid DIP rounding issues.
This is a local fix to the larger problem of scaled bitmaps rendered in
ImageViews at non-integral DIP scaling factors (125%, 133%, 175%, etc.)

It solves the specific bug, but does not solve the larger problem of the
ways we scale bitmaps and Views canvases being sometimes different.

More info is available on the issue cited.

I am looking for feedback - if this is an acceptable fix for right now
or if we should drive towards a more comprehensive approach.

Probably both.

Bug: 1094566
Change-Id: I024e216542cee057cd4198a5a65344ffa0cd999d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2306751
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791060}