[omnibox] Remove verbatim match from remote ZeroSuggest on Desktop
We've always returned the current page's URL as a verbatim default match
for ZeroSuggest (with the exception of on the NTP).

However, we don't need this anymore on Desktop.

This CL preserves the most-visited mode's behavior, but removes the
verbatim match from all remotely fetched (from the suggest server)
ZeroSuggest results on Desktop.

We leave Android and iOS alone. Android needs the verbatim match for
the Search Ready Omnibox edit widget. iOS we're leaving alone because
there's no strong reason to change it.

Bug: 1095843, 1106096
Change-Id: Iac6c1ee401a93cda05df7cc96d5e176e3d36f3ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308944
Reviewed-by: Ender 💬 = 🕐 (ping me for faster reviews) <ender@google.com>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791063}