Fix heap corruption during VK show/hide.
In VK APIs, we register callbacks with input pane whenever show/hide
APIs are called. These callbacks are called by the OS when VK is about
to be shown or hidden. Sometimes, the callbacks from the OS comes after
the owner of the COM objects get destroyed. This submits a task that
has a BindState that contains a tuple of WeakPtr to the owner of |this|
object and the functor. For some reason(may be due to timing) that has
already been destroyed. The COM objects get released by the OS after
the Worker thread COM delegate task runner gets flushed in
|WorkerThreadCOMDelegate::GetWorkFromWindowsMessageQueue|. This patch
adds checks to prevent posting of tasks if the COM objects are invalid.
It also adds code to de-register the callbacks during destruction of
the VK ref counted object.

Bug: 1104227

Change-Id: Icdd7dd309eac941f23eed88184db34d77bd28571
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2311721
Commit-Queue: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Yohei Yukawa <yukawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791107}