[files-progress] Add new panel styles for transfer details
Add the following attributes to the existing xf-panels to flag this
new feature (transfer details) via CSS style:

- detailed-panel:
  All new panels (including summary panel) indicated by this attribute.
- detailed-summary:
  Indicates summary panel.
- data-category(expanded|collapsed):
  Indicating panel collapsed or expanded status.

Currently only the expand/collapse button keeps this attribute(status)
but the CL sets this attribute in the panel level as well. It is used
by summary panel text to expand and collapse it along with button. The
|textDiv| is added to handle the expand/collapse status of panels.

Update existing colors to those in files-ng spec. These color values
are identical to the non-files-ng colors, so no visible color change
for the non-files-ng case.

Bug: 953308
Change-Id: Ie456cc31594dec802757729be50e0434653192bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2307551
Commit-Queue: Omid Tourzan <oto@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791138}