[arraybuffer] Create backing store also when data is nullptr
In the constructor of ArrayBufferContents, we did not create a
V8::BackingStore if an byte array was provided, i.e. byte == nullptr.
Without BackingStore, however, it was not possible to create a
V8::ArrayBuffer for this ArrayBufferContents, as V8 expected that a
BackingStore always exists. With this CL we always create a
BackingStore, even though the data is nullptr.

R=haraken@chromium.org, ulan@chromium.org

Bug: chromium:1104580
Change-Id: I4a05de503f5916a9491dbe4595f699c0e834bb10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302694
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Ulan Degenbaev <ulan@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791184}