Add SCTReportingService and E2E tests
This adds SCT auditing enable/disable interfaces to NetworkContext to
control auditing behavior on a per-NetworkContext basis, and a new
SCTReportingService to monitor changes to the Safe Browsing Extended
Reporting status per-profile (to gate SCT audit collection to users
opting in to Extended Reporting). The SCTReportingService also receives
the notifications about new SCTAuditReports sent to the
NetworkContextClient.

As this makes entries in the SCTAuditingCache visible to the browser,
this also adds end-to-end tests covering the cache behavior.

Bug: 1082860
Change-Id: I98922bbdc359cc641b920185fd0b4ccd8aea9eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252569
Auto-Submit: Christopher Thompson <cthomp@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791210}