[Autofill Assistant] Changing progress bar icons must keep state
Before this change the state of the progress bar got lost when changing
icons. This means that icons and lines marked as "done" would become
unmarked again and the currently active (i.e. animated) icon would stop
being active.

This generally is only a problem for the first icon when starting with
the proper experiment. The flow would set the active step to 0,
then change the icons and thus lose the state.

Additionally, it was not possible to stop using the new progress
bar without specifying icons, as the action would fail.

Bug: b/161962005
Change-Id: I80592cb6215a94a95dd67fe66a7d5d494e0f3c28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2315883
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Luca Hunkeler <hluca@google.com>
Cr-Commit-Position: refs/heads/master@{#791270}