Test Ash smart card login signature failure
Provide test coverage for the scenario when the signature operation,
triggered during the challenge-response login (a.k.a. smart card based
login), fails.

This corresponds to the real-world case when a smart card user tries to
log into the system but the smart card is broken or is ejected during
the login.

Bug: 1090941
Change-Id: Ida39fa52385905643bbeb3728dbbc2f49c527195
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2315676
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791276}