[iOS] Do not display SSL interstitial for a WebState not in a WebStateList
When a WebState isn't in a WebStateList, it doesn't have tab
helpers. WebStates used to update the reading list in the
background aren't in any WebStateList. When such a WebState
encounters an SSL error, it currently attempts to display
an error interstitial.

With committed interstitials, IOSBlockingPageTabHelper
is used to display SSL errors, but since tab helpers aren't
created for WebStates that aren't in a WebStateList, trying
to use an IOSBlockingPageTabHelper results in a crash.

Without committed interstitials, an error interstitial is
created and displayed in the WebState, but is never seen by
the user since the WebState itself is not visible (and will
never be visible).

This CL stops attempting to display an SSL error interstitial
for a WebState that isn't in the WebStateList.

(cherry picked from commit 8ef27fbdc7e40cf43933241a27029b60198f06f3)

Bug: 1107534
Change-Id: Iae88f30a024dd770f59b15b1aa3fc6f2ea23f06a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308910
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790508}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316318
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#893}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}