Use extended process attributes to associate job at startup.
On Windows 10 we can ask Windows to associate a process with a job
at startup. This reduces the time during which the process handle
can be used to launch more processes.

Prior to Windows 10 we retain the existing call to AssignProcessToJobObject.

Existing tests should cover this change. I have manually verified that
processes are associated with Jobs using SysInternals ProcExp.

Bug: 1050359
Change-Id: I339e42be7f30cf0e59554e89f4855dcd3af4d499
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308916
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791320}