Remove CPMC from RenderWidgetHost InputEvent observers
Fix: remove ChromePasswordManagerClient from the
InputEventObserver list on WebContentsDestroyed() and
RenderFrameDeleted().

Problem summary: We get an input event when the instance of
ChromePasswordManagerClient is invalid. Reading through the code,
this probably happens because the CPMC is bound to WebContents,
which does get destroyed, but we do not unregister CPMC from the
InputEvent observer list.

Notes:
* This is a speculative fix because I could not reproduce the crash
  locally.
* I am also not sure how to add a test for this condition but I'd
  be happy to add one with guidance.

Bug: 1104919
Change-Id: I3077db25343051c5f2a1e3e905e8abec53556f57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2309814
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791358}