[bluetooth] Enumeration of serial devices
For enumeration of serial devices, it is necessary to register
an observer using an adapter instance. With the adapter it is
also possible to use GetDevices() to get a list of the current
discovered devices which should allow serial devices to show up
in the serial device chooser.

Bug: 1043300
Change-Id: I2bc0263f0f2baaacbdc86b6d97034017dff61412
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2291316
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Martin Barbella <mbarbella@chromium.org>
Commit-Queue: Vicky Min <vickymin@google.com>
Cr-Commit-Position: refs/heads/master@{#791364}