[WPT test] Update text.measurement expectation
After verifying the spec, I learned that
measureText('A  B').width != measureText('A B').width. The spaces do not
suppose to collapse:
-> Replace all ASCII whitespace in text with U+0020 SPACE characters.
-> Then the 'white-space' property set to 'pre'.
-> Then Let result be an array constructed by iterating over each glyph
in the inline box from left to right (if any), adding to the array, for
each glyph, the shape of the glyph as it is in the inline box,
positioned on a coordinate space using CSS pixels with its origin is at
the anchor point.

Update the test expectation accordingly.

Bug: 1066953

Change-Id: Ifb9885815f3cbf3d46432107abf16af6ed5b4cf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2306575
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Commit-Queue: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791412}