[iOS] Don't reset the alert coordinator when cancel
It seems that resetting the alert coordinator is actually dismissing the
ReadingList ViewController.
This is not needed as nothing is relying on the coordinator to be nil,
so removing the nil assignment is working.

Fixed: 1105852
Change-Id: I1cf8c38e6f6b8d1f8456e6b2ed439e38c44db275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2316204
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#791453}